Project is now under GNU GPL3 License
I should have done this a long time ago, I was too focused on functionality and I forgot to state to users what's the real intention of this software, so here it go...

The whole idea is to learn and improve my programming skills, making something complex, fun and useful at some extent not only for myself. In the process, learn how to use modern C++ features and experiment with best pratices, ultimately becoming a better programmer everyday.  There are plenty to experiment in this project, like software engineering, CPU, GPU and memory access optimizations, editor and game modules, scripting, portability, graphics, physics, sound, player input, system/application threads, and later on the creation of a game using the features developed along the way.

I would absolutely love to have other peoples opinion on the code, have suggestions on how to make it better and have other developers working on it with me.

My goal is to have a portable and easy to use game engine, which people can easily extend for any application and field they want, make it extremely simple to get started with and at the same time be very customizable/powerful as the user get use with the framework. It would be great seeing people using it to start any 3D rendering they need, along side Blender and other open source softwares.

One day I want to hand it to my children and say "do wathever you like there, here are some examples, have fun!".